feat: Add discount information to payment plans
Introduces a 'discount' field to both CALL_PLANS and CHAT_PLANS constants, as well as to the Plan interface in types.ts. This allows for displaying discount percentages on the UI, enhancing user visibility of savings on various plan options. The PlanCard and Listeners components have been updated to render this discount information.